requirements.txt
-i https://pypi.org/simple
certifi==2018.4.16
chardet==3.0.4
idna==2.7
oauthlib==2.1.0
python-dotenv==0.8.2
requests-oauthlib==1.0.0
requests==2.19.1
twython==3.7.0
urllib3==1.23
